# Improve `TaggedLogging::Formatter` performance